rename YamlSetsResolver to SetsParameterResolver
