+++ packages/set-config-resolver/src/ConfigResolver.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace Symplify\SetConfigResolver;
 
 use Symfony\Component\Console\Input\InputInterface;
+use Symplify\SetConfigResolver\Config\SetsParameterResolver;
 use Symplify\SetConfigResolver\Console\Option\OptionName;
 use Symplify\SetConfigResolver\Console\OptionValueResolver;
-use Symplify\SetConfigResolver\Yaml\YamlSetsResolver;
 use Symplify\SmartFileSystem\Exception\FileNotFoundException;
 use Symplify\SmartFileSystem\SmartFileInfo;
 
@@ -19,7 +19,7 @@ final class ConfigResolver
     private $firstResolverConfig;
 
     /**
-     * @var YamlSetsResolver
+     * @var \Symplify\SetConfigResolver\Config\SetsParameterResolver
      */
     private $yamlSetsResolver;
 
@@ -36,7 +36,7 @@ final class ConfigResolver
     public function __construct()
     {
         $this->setResolver = new SetResolver();
-        $this->yamlSetsResolver = new YamlSetsResolver();
+        $this->yamlSetsResolver = new SetsParameterResolver();
         $this->optionValueResolver = new OptionValueResolver();
     }
